user rel department yozadigan qilindi

+++ models/Users.php
@@ -4,7 +4,9 @@ namespace app\models;
 
 use app\modules\admin\models\AuthAssignment;
 use app\modules\admin\models\RedirectUrlList;
+use app\modules\hr\models\HrEmployeeRelPosition;
 use app\modules\hr\models\HrEmployeeRelUsers;
+use app\modules\hr\models\UsersRelationHrDepartments;
 use Yii;
 
 /**
@@ -267,11 +269,15 @@ class Users extends BaseModel implements \yii\web\IdentityInterface
                     'message' => Yii::t('app', 'Hr employee id required')
                 ];
 
-            if ($this->isUpdate && $response['status']){
+            if ($this->isUpdate && $response['status'] && $this->id){
                 HrEmployeeRelUsers::deleteAll(['user_id' => $this->id]);
                 AuthAssignment::deleteAll(['user_id' => $this->id]);
+                UsersRelationHrDepartments::deleteAll(['user_id' => $this->id]);
             }
 
+            /**
+             * Hodimga oldin foydalanuvchi biriktirilganligini tekshiradi
+             */
             if ($response['status']){
                 $existsHrEmployeeUser = HrEmployeeRelUsers::checkExistsHrEmployeeUser($this->hr_employee_id);
                 if ($existsHrEmployeeUser)
@@ -289,6 +295,9 @@ class Users extends BaseModel implements \yii\web\IdentityInterface
                         'errors' => $this->getErrors()
                     ];
 
+            /**
+             * Hodimga foydalanuvchi biriktiradi
+             */
             if ($response['status']){
                 $hrEmployeeRelUsers = new HrEmployeeRelUsers([
                     'hr_employee_id' => $this->hr_employee_id,
@@ -302,6 +311,9 @@ class Users extends BaseModel implements \yii\web\IdentityInterface
                     ];
             }
 
+            /**
+             * Rollarni saqlaydi
+             */
             if ($response['status'] && !empty($this->roles)){
                 foreach ($this->roles as $key => $role)
                 {
@@ -318,6 +330,41 @@ class Users extends BaseModel implements \yii\web\IdentityInterface
                         ];
                         break;
                     }
+                }
+            }
+
+            /**
+             * Foydalanuvchiga bo'lim va tashkilotni biriktirib qo'yadi
+             */
+            if ($response['status']){
+                $relEmployeeDepartment = HrEmployeeRelPosition::findOne(["hr_employee_id" => $this->hr_employee_id, 'status_id' => BaseModel::STATUS_ACTIVE]);
+
+                if(!empty($relEmployeeDepartment) && $relEmployeeDepartment->hr_organisation_id){
+                    $userRelOrganisations = new UsersRelationHrDepartments([
+                        "user_id" => $this->id,
+                        "hr_department_id" => $relEmployeeDepartment->hr_organisation_id,
+                        "is_root" => UsersRelationHrDepartments::ROOT,
+                    ]);
+                    if (!$userRelOrganisations->save())
+                        $response = [
+                            'status' => false,
+                            'message' => Yii::t('app', 'User rel organisations not saved'),
+                            'errors' => $userRelOrganisations->getErrors()
+                        ];
+                }
+
+                if($response['status'] && !empty($relEmployeeDepartment) && $relEmployeeDepartment->hr_department_id){
+                    $userRelDepartments = new UsersRelationHrDepartments([
+                        "user_id" => $this->id,
+                        "hr_department_id" => $relEmployeeDepartment->hr_department_id,
+                        "is_root" => UsersRelationHrDepartments::NOT_ROOT,
+                    ]);
+                    if (!$userRelDepartments->save())
+                        $response = [
+                            'status' => false,
+                            'message' => Yii::t('app', 'User rel department not saved'),
+                            'errors' => $userRelDepartments->getErrors()
+                        ];
                 }
             }
